==> 02_stateanalysis_markdown.Rmd <==
---
title: "Food Service Workers"
author: "Aaron Kessler"
date: "3/18/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(janitor)
library(readxl)
library(writexl)


#import data for all states, all occupations
rawdata <- read_excel("raw_data/oesm18st/state_M2018_dl.xlsx")

#clean names
states_allrecs <- rawdata %>% 
  clean_names()


#cut down to only the measures we'll use here:

# -- tot_emp: Estimated total employment rounded to the nearest 10 (excludes self-employed)
# -- jobs_1000: The number of jobs (employment) in the given occupation per 1,000 jobs in the given area 
# -- h_mean: Mean hourly wage
# -- a_mean: Mean annual wage 
# -- annual: Contains "TRUE" if only the annual wages are released
# -- hourly:	Contains "TRUE" if only the hourly wages are released

states_allrecs <- states_allrecs %>% 
  select(
    area,
    st,
    state,
    occ_code,
    occ_title,
    occ_group,
    tot_emp,
    jobs_1000,
    h_mean,
    a_mean
  )


glimpse(states_allrecs)

#format columns
states_allrecs <- states_allrecs %>% 
  mutate(
    tot_emp = as.numeric(tot_emp),
    jobs_1000 = as.numeric(jobs_1000),
    h_mean = as.numeric(h_mean),
    a_mean = as.numeric(a_mean)
  )

#limit to just the 50 states and DC
states_allrecs <- states_allrecs %>% 
  filter(st %in% c("GU",
  "VI",
  "PR"))


```

*Data source: BLS Occupational Employment Statistics (2018)* 
*https://www.bls.gov/oes/tables.htm*  
  
## Filtering to just food prep and server workers

```{r }

foodserv_all <- states_allrecs %>% 
  filter(str_starts(occ_code, "35-"))

```

Check to see if any our missing or redacted
```{r }
 
foodserv_all %>% 
  filter(is.na(tot_emp))

```

Just a handful. No worries then, we can work with that.    
  
To keep things simple to start, we'll just look at the overarching total from the major category record for each state.

```{r}

foodserv_totalsonly <- foodserv_all %>% 
  filter(occ_code == "35-0000")

foodserv_totalsonly

foodserv_all %>% 
  count(state, st)



```

```{r}




```

```{r}




```

```{r}




```

```{r}




```

```{r}




```

```{r}




```

```{r}




```

```{r}




```

```{r}




```
